requirements.txt
aiobotocore==2.25.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aioitertools==0.12.0
aiosignal==1.4.0
appdirs==1.4.4
attrs==25.4.0
botocore==1.40.61
-e git+ssh://git@github.com/CCI-GU-Sweden/CCIPyUtils.git@f8500b2766a0e43bb587129df7c4980f80113b85#egg=ccipy_omero&subdirectory=ccipy-omero
-e git+ssh://git@github.com/CCI-GU-Sweden/CCIPyUtils.git@f8500b2766a0e43bb587129df7c4980f80113b85#egg=ccipy_utils&subdirectory=ccipy-utils
-e git+ssh://git@github.com/CCI-GU-Sweden/CCIPyUtils.git@f8500b2766a0e43bb587129df7c4980f80113b85#egg=ccipy_yolo_wrapper&subdirectory=ccipy-yolo-utils
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.2
contourpy==1.3.3
crc32c==2.8
cuda-toolkit==12.2.0
cycler==0.12.1
dask==2025.10.0
donfig==0.8.1.post1
filelock==3.20.0
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.10.0
future==1.0.0
idna==3.11
imagecodecs==2025.8.2
ImageIO==2.37.2
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.9
lazy_loader==0.4
locket==1.0.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
networkx==3.5
nodeenv==1.9.1
numcodecs==0.16.3
numpy==2.2.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.1.105
ome-zarr==0.12.2
omero-py==5.21.2
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
partd==1.4.2
pillow==12.0.0
polars==1.35.2
polars-runtime-32==1.35.2
portalocker==3.2.0
propcache==0.4.1
psutil==7.1.3
Pygments==2.19.2
pyparsing==3.2.5
pyright==1.1.407
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
s3fs==2025.10.0
scikit-image==0.25.2
scipy==1.16.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
tifffile==2025.10.16
toolz==1.1.0
torch==2.4.1
torchvision==0.19.1
triton==3.0.0
typer==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.233
ultralytics-thop==2.0.18
urllib3==2.5.0
wrapt==1.17.3
xmltodict==1.0.2
yarl==1.22.0
zarr==3.1.3
zeroc-ice @ https://github.com/glencoesoftware/zeroc-ice-py-linux-x86_64/releases/download/20240202/zeroc_ice-3.6.5-cp312-cp312-manylinux_2_28_x86_64.whl#sha256=96fb9066912c52f2503e9f9207f98d51de79c475d19ebd3157aae7bc522b5826
